api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pymongo==4.6.0
python-multipart==0.0.6
pandas==2.1.3